IAT-524: Refactor core services to flat directory structure and names ending with .service.
