+++ src/app/core/app-info.service/app-info.ts
