+++ src/app/core/core.module.ts
@@ -1,16 +1,16 @@
 import {NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
 
-import {AlertService} from "./services/alert/alert.service";
-import {Logger} from "./services/logger/logger.service";
+import {AlertService} from "./alert.service/alert.service";
+import {Logger} from "./logger.service/logger.service";
 import {ItemTypeService} from "../item/services/item-type/item-type.service";
-import {UserService} from "./services/user/user.service";
+import {UserService} from "./user.service/user.service";
 import {HttpModule} from "@angular/http";
 import {AlertModule} from "ngx-bootstrap";
-import {BusyService} from "./services/busy/busy.service";
-import {BusyOverlayComponent} from "./services/busy/busy-overlay.component";
-import {AppInfoService} from "./services/app-info/app-info.service";
-import {HttpUtility} from "./services/http-utility/http-utility";
+import {BusyService} from "./busy.service/busy.service";
+import {BusyOverlayComponent} from "./busy.service/busy-overlay.component";
+import {AppInfoService} from "./app-info.service/app-info.service";
+import {HttpUtility} from "./http-utility.service/http-utility";
 
 @NgModule({
   declarations: [
