+++ src/app/item/components/item-dashboard/create-item-select-type/create-item-select-type.component.spec.ts
@@ -1,11 +1,11 @@
 import {async, TestBed} from "@angular/core/testing";
 import {HttpModule} from "@angular/http";
 import {CreateItemSelectTypeComponent} from "./create-item-select-type.component";
-import {AppInfoService} from "../../../../core/services/app-info/app-info.service";
+import {AppInfoService} from "../../../../core/app-info.service/app-info.service";
 import {RouterTestingModule} from "@angular/router/testing";
-import {Logger} from "../../../../core/services/logger/logger.service";
+import {Logger} from "../../../../core/logger.service/logger.service";
 import {ItemTypeService} from "../../../services/item-type/item-type.service";
-import {AlertService} from "../../../../core/services/alert/alert.service";
+import {AlertService} from "../../../../core/alert.service/alert.service";
 import {ToastyConfig, ToastyService} from "ng2-toasty";
 import {CoreModule} from "../../../../core/core.module";
 import {ItemService} from "../../../services/item/item.service";
