+++ src/app/item/components/item-dashboard/item-dashboard.component.spec.ts
@@ -2,8 +2,8 @@ import {async, TestBed} from "@angular/core/testing";
 import {APP_BASE_HREF} from "@angular/common";
 
 import {ItemDashboardComponent} from "./item-dashboard.component";
-import {Logger} from "../../../core/services/logger/logger.service";
-import {AlertService} from "../../../core/services/alert/alert.service";
+import {Logger} from "../../../core/logger.service/logger.service";
+import {AlertService} from "../../../core/alert.service/alert.service";
 import {AppModule} from "../../../root/app.module";
 
 describe('ItemDashboardComponent', () => {
