+++ src/app/item/components/item/item-auto-save/item-auto-save.component.spec.ts
@@ -1,7 +1,7 @@
 import {async, ComponentFixture, TestBed} from "@angular/core/testing";
 
 import {ItemAutoSaveComponent} from "./item-auto-save.component";
-import {Logger} from "../../../../core/services/logger/logger.service";
+import {Logger} from "../../../../core/logger.service/logger.service";
 import {ItemService} from "../../../services/item/item.service";
 import {CoreModule} from "../../../../core/core.module";
 import {ToastyConfig, ToastyService} from "ng2-toasty";
