+++ src/app/item/components/item/item-details/item-history/item-history.component.spec.ts
@@ -3,8 +3,8 @@ import {APP_BASE_HREF} from "@angular/common";
 
 import {ItemHistoryComponent} from "./item-history.component";
 import {AppModule} from "../../../../../root/app.module";
-import {Logger} from "../../../../../core/services/logger/logger.service";
-import {AlertService} from "../../../../../core/services/alert/alert.service";
+import {Logger} from "../../../../../core/logger.service/logger.service";
+import {AlertService} from "../../../../../core/alert.service/alert.service";
 import {ItemService} from "../../../../services/item/item.service";
 
 describe('ItemHistoryComponent', () => {
