+++ src/app/item/components/item/item-types/wer-item/wer-item.component.spec.ts
@@ -1,6 +1,6 @@
 import {async, ComponentFixture, TestBed} from "@angular/core/testing";
 import {WerItemComponent} from "./wer-item.component";
-import {Logger} from "../../../../../core/services/logger/logger.service";
+import {Logger} from "../../../../../core/logger.service/logger.service";
 import {AppModule} from "../../../../../root/app.module";
 import {APP_BASE_HREF} from "@angular/common";
