+++ src/app/item/components/item/item.component.spec.ts
@@ -1,8 +1,8 @@
 import {async, ComponentFixture, TestBed} from "@angular/core/testing";
 import {APP_BASE_HREF} from "@angular/common";
 import {AppModule} from "../../../root/app.module";
-import {Logger} from "../../../core/services/logger/logger.service";
-import {AlertService} from "../../../core/services/alert/alert.service";
+import {Logger} from "../../../core/logger.service/logger.service";
+import {AlertService} from "../../../core/alert.service/alert.service";
 import {ItemComponent} from "./item.component";
 
 describe('ItemComponent', () => {
