+++ src/app/root/components/footer/footer.component.spec.ts
@@ -2,12 +2,12 @@ import {async, TestBed} from "@angular/core/testing";
 
 import {FooterComponent} from "./footer.component";
 import {HttpModule} from "@angular/http";
-import {AppInfoService} from "../../../core/services/app-info/app-info.service";
-import {AlertService} from "../../../core/services/alert/alert.service";
+import {AppInfoService} from "../../../core/app-info.service/app-info.service";
+import {AlertService} from "../../../core/alert.service/alert.service";
 import {ToastyConfig, ToastyService} from "ng2-toasty";
-import {Logger} from "../../../core/services/logger/logger.service";
-import {HttpUtility} from "../../../core/services/http-utility/http-utility";
-import {BusyService} from "../../../core/services/busy/busy.service";
+import {Logger} from "../../../core/logger.service/logger.service";
+import {HttpUtility} from "../../../core/http-utility.service/http-utility";
+import {BusyService} from "../../../core/busy.service/busy.service";
 
 describe('FooterComponent', () => {
   beforeEach(async(() => {
