+++ src/app/root/components/header/header.component.spec.ts
@@ -4,12 +4,12 @@ import {HeaderComponent} from "./header.component";
 import {BsDropdownModule} from "ngx-bootstrap/dropdown";
 import {ModalModule} from "ngx-bootstrap/modal";
 import {CollapseDirective} from "ngx-bootstrap/collapse";
-import {Logger} from "../../../core/services/logger/logger.service";
-import {UserService} from "../../../core/services/user/user.service";
-import {AlertService} from "../../../core/services/alert/alert.service";
+import {Logger} from "../../../core/logger.service/logger.service";
+import {UserService} from "../../../core/user.service/user.service";
+import {AlertService} from "../../../core/alert.service/alert.service";
 import {ToastyConfig, ToastyService} from "ng2-toasty";
-import {HttpUtility} from "../../../core/services/http-utility/http-utility";
-import {BusyService} from "../../../core/services/busy/busy.service";
+import {HttpUtility} from "../../../core/http-utility.service/http-utility";
+import {BusyService} from "../../../core/busy.service/busy.service";
 
 describe('HeaderComponent', () => {
   beforeEach(async(() => {
